requirements.txt
fastapi
uvicorn
python-dotenv
spotipy
openai
jinja2
python-multipart
lyricsgenius 